backend/requirements.txt
# FastAPI Backend - Python Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database ORM
sqlmodel==0.0.14
psycopg[binary]==3.2.13

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # Pin compatible version for passlib

# Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# CORS
python-multipart==0.0.6
